Split server into async and sync, implemented some server async.
