+++ src/lib.rs
@@ -63,7 +63,6 @@ extern crate bitflags;
 extern crate test;
 
 pub use self::client::{Client, ClientBuilder};
-pub use self::server::Server;
 pub use self::dataframe::DataFrame;
 pub use self::message::{Message, OwnedMessage};
 pub use self::stream::Stream;
